CFP2-130: Extract rental-units and tenancies table as a separate components
