+++ src/components/rental-unit/overview/rental-units-overview.tsx
@@ -1,19 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import {
-    Grid,
-    IconButton,
-    makeStyles,
-    Paper,
-    Table,
-    TableBody,
-    TableCell,
-    TableContainer,
-    TableHead,
-    TableRow,
-    Typography,
-} from '@material-ui/core';
-import { getDisplayNameOfRentalUnit } from '../model/rental-unit.model';
-import { Link } from 'react-router-dom';
+import { Grid, Typography } from '@material-ui/core';
 import routes from '../../../routes/route-constants';
 import { useAppDispatch, useAppSelector } from '../../../hooks/store.hooks';
 import {
@@ -23,24 +9,13 @@ import {
     selectTenancies,
 } from '../../../store/selectors';
 import { deleteRentalUnit } from '../../../store/slices/rental-units.slice';
-import EditOutlinedIcon from '@material-ui/icons/EditOutlined';
-import DeleteOutlineIcon from '@material-ui/icons/DeleteOutline';
-import { useDeletePrompt } from '../../../hooks/ui.hooks';
-import DeletePrompt from '../../ui/delete-prompt/delete-prompt';
-import { getTenantsOfTenancy } from '../../tenancy/model/tenancy.model';
-import { parseISO } from 'date-fns';
 import { deleteTenancy } from '../../../store/slices/tenancies.slice';
 import SearchHeader from '../../ui/search-header/search-header';
 import { useHistory } from 'react-router';
-
-const useStyles = makeStyles({
-    table: {
-        width: '100%',
-    },
-});
+import { RentalUnitsTable } from './rental-units-table';
+import { stylingConstants } from '../../../theme/shared-styles';
 
 export const RentalUnitsOverview = ({ disableCreate }: { disableCreate: boolean }) => {
-    const cssClasses = useStyles();
     const history = useHistory();
 
     const dispatch = useAppDispatch();
@@ -50,9 +25,6 @@ export const RentalUnitsOverview = ({ disableCreate }: { disableCreate: boolean 
     const tenancies = useAppSelector(selectTenancies).filter((ten) => ten.propertyId === property?.id);
     const tenants = useAppSelector(selectPersonsTenants);
 
-    const { deletePromptOpen, entityToDelete, handleOpenDeletePrompt, handleCancelDelete } =
-        useDeletePrompt();
-
     const [searchResult, setSearchResult] = useState(rentalUnits);
 
     // TODO: Fix 'Maximum update depth exceeded' error
@@ -60,11 +32,11 @@ export const RentalUnitsOverview = ({ disableCreate }: { disableCreate: boolean 
         setSearchResult(rentalUnits);
     }, [rentalUnits]);
 
-    const handleDelete = () => {
-        dispatch(deleteRentalUnit(entityToDelete));
+    const handleDelete = (ruId: string) => {
+        dispatch(deleteRentalUnit(ruId));
         // TODO: Combine these store actions within the store (transparent)
         tenancies
-            ?.filter((ten) => ten.rentalUnitId === entityToDelete)
+            ?.filter((ten) => ten.rentalUnitId === ruId)
             ?.forEach((ten) => {
                 dispatch(deleteTenancy(ten.id));
             });
@@ -74,24 +46,16 @@ export const RentalUnitsOverview = ({ disableCreate }: { disableCreate: boolean 
         history.push(routes.RENTAL_UNITS_CREATE);
     };
 
-    const getTenantsDesc = (ruId: string) => {
-        const ts = getTenantsOfTenancy(
-            // TODO reuse filters and finds
-            tenancies
-                ?.filter((t) => !t.endOfContract || parseISO(t.endOfContract) > new Date())
-                ?.find((ten) => ten.rentalUnitId === ruId),
-            tenants
-        );
-
-        if (!ts.length) {
-            return 'Vacancy';
-        }
-
-        return ts.join(', ');
-    };
-
     return (
-        <>
+        <Grid
+            item
+            container
+            xs={12}
+            sm={6}
+            spacing={stylingConstants.gridSpacing}
+            alignItems={'center'}
+            alignContent={'flex-start'}
+        >
             <Grid item xs={12}>
                 <Typography variant={'h6'} component={'h2'}>
                     Associated Rental Units
@@ -107,52 +71,13 @@ export const RentalUnitsOverview = ({ disableCreate }: { disableCreate: boolean 
                 wrapAtMd={true}
             />
             <Grid item xs={12}>
-                <DeletePrompt
-                    open={deletePromptOpen}
-                    title={'Delete Rental Unit?'}
-                    description={
-                        'Are you sure you want to delete this rental unit? This will also delete all linked tenancies!'
-                    }
-                    handleClose={handleCancelDelete}
-                    handleDeletion={handleDelete}
+                <RentalUnitsTable
+                    tenancies={tenancies}
+                    tenants={tenants}
+                    handleDelete={handleDelete}
+                    searchResult={searchResult}
                 />
-                {/* TODO: Check if it makes sense to extract table as component */}
-                <TableContainer component={Paper}>
-                    <Table className={cssClasses.table}>
-                        <TableHead>
-                            <TableRow>
-                                <TableCell>Actions</TableCell>
-                                <TableCell align="right">Description</TableCell>
-                                <TableCell align="right">Current Tenant</TableCell>
-                            </TableRow>
-                        </TableHead>
-                        <TableBody>
-                            {searchResult?.map((ru) => (
-                                <TableRow key={ru.id}>
-                                    <TableCell>
-                                        <IconButton
-                                            aria-label={'edit'}
-                                            component={Link}
-                                            to={routes.getRentalUnitDetailRouteById(ru.id)}
-                                        >
-                                            <EditOutlinedIcon />
-                                        </IconButton>
-                                        <IconButton
-                                            aria-label={'delete'}
-                                            onClick={() => handleOpenDeletePrompt(ru.id)}
-                                        >
-                                            <DeleteOutlineIcon color={'error'} />
-                                        </IconButton>
-                                    </TableCell>
-                                    <TableCell align="right">{getDisplayNameOfRentalUnit(ru)}</TableCell>
-                                    {/* // TODO: Make font bold and red (error) if it is a Vacancy */}
-                                    <TableCell align="right">{getTenantsDesc(ru.id)}</TableCell>
-                                </TableRow>
-                            ))}
-                        </TableBody>
-                    </Table>
-                </TableContainer>
             </Grid>
-        </>
+        </Grid>
     );
 };
